Move JSONL record reading out of the upload module
iter_jsonl, count_jsonl, and chunks are needed by every consumer of
materialized record files, not just the Qdrant uploader. Moving them to
a neutral records module lets upcoming consumers (the embed step) avoid
importing Qdrant code. The names are re-exported from upload, so
existing imports keep working.
